requirement.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
boto3==1.26.165
stripe==5.5.0
email-validator==2.0.0.post2